+++ backtrace-library/src/androidTest/java/backtraceio/library/models/SkipExceptionHandler.java
@@ -1,0 +1,10 @@
+package backtraceio.library.models;
+
+public class SkipExceptionHandler implements Thread.UncaughtExceptionHandler {
+    @Override
+    public void uncaughtException(Thread t, Throwable e) {
+        System.err.println("Exception caught in thread " + t.getName() + ":");
+        e.printStackTrace();
+        System.err.println("Exception skipped");
+    }
+}