+++ backtrace-library/src/main/java/backtraceio/library/BacktraceDatabase.java
@@ -12,6 +12,7 @@ import java.util.concurrent.CountDownLatch;
 import backtraceio.library.base.BacktraceBase;
 import backtraceio.library.breadcrumbs.BacktraceBreadcrumbs;
 import backtraceio.library.common.FileHelper;
+import backtraceio.library.common.TypeHelper;
 import backtraceio.library.enums.UnwindingMode;
 import backtraceio.library.enums.database.RetryBehavior;
 import backtraceio.library.events.OnServerResponseEventListener;
@@ -39,7 +40,6 @@ public class BacktraceDatabase implements Database {
 
     private final String _crashpadHandlerName = "/libcrashpad_handler.so";
     private final String _crashpadDatabasePathPrefix = "/crashpad";
-
     private static boolean _timerBackgroundWork = false;
     private static Timer _timer;
     private transient final String LOG_TAG = BacktraceDatabase.class.getSimpleName();
@@ -51,6 +51,7 @@ public class BacktraceDatabase implements Database {
     private boolean _enable = false;
     private Breadcrumbs breadcrumbs;
 
+    private boolean _enabledNativeIntegration = false;
     /**
      * Add attributes to native reports
      *
@@ -168,7 +169,7 @@ public class BacktraceDatabase implements Database {
     public Boolean setupNativeIntegration(BacktraceBase client, BacktraceCredentials credentials,
                                           boolean enableClientSideUnwinding, UnwindingMode unwindingMode) {
         // avoid initialization when database doesn't exist
-        if (getSettings() == null) {
+        if (_enable == false || getSettings() == null) {
             return false;
         }
         String minidumpSubmissionUrl = credentials.getMinidumpSubmissionUrl().toString();
@@ -203,7 +204,7 @@ public class BacktraceDatabase implements Database {
         File crashHandlerDir = new File(databasePath);
         crashHandlerDir.mkdir();
 
-        Boolean initialized = initialize(
+        _enabledNativeIntegration = initialize(
                 minidumpSubmissionUrl,
                 databasePath,
                 handlerPath,
@@ -214,12 +215,12 @@ public class BacktraceDatabase implements Database {
                 unwindingMode
         );
 
-        if (initialized && this.breadcrumbs.isEnabled()) {
+        if (_enabledNativeIntegration && this.breadcrumbs.isEnabled()) {
             this.breadcrumbs.setOnSuccessfulBreadcrumbAddEventListener(breadcrumbId -> {
                 this.addAttribute("breadcrumbs.lastId", Long.toString((breadcrumbId)));
             });
         }
-        return initialized;
+        return _enabledNativeIntegration;
     }
 
     /**
@@ -228,11 +229,28 @@ public class BacktraceDatabase implements Database {
     @Override
     public void disableNativeIntegration() {
         disable();
+        this._enabledNativeIntegration = false;
     }
 
     @Override
     public Breadcrumbs getBreadcrumbs() {
         return this.breadcrumbs;
+    }
+
+    public Boolean addNativeAttribute(String key, Object value) {
+        if (!_enabledNativeIntegration) {
+            return false;
+        }
+
+        if (key == null || value == null) {
+            return false;
+        }
+        Class type = value.getClass();
+        if (!TypeHelper.isPrimitiveOrPrimitiveWrapperOrString(type)) {
+            return false;
+        }
+        addAttribute(key, value.toString());
+        return true;
     }
 
     public void start() {
@@ -410,6 +428,10 @@ public class BacktraceDatabase implements Database {
 
     public void delete(BacktraceDatabaseRecord record) {
         if (this.backtraceDatabaseContext == null) {
+            return;
+        }
+
+        if (record == null){
             return;
         }
         this.backtraceDatabaseContext.delete(record);